Реализация разделов: news; history; documents; projects. И их HTTP методов v.2

+++ src/main/java/com/example/rezh/services/DocumentService.java
@@ -3,6 +3,7 @@ package com.example.rezh.services;
 
 import com.example.rezh.entities.DocumentEntity;
 import com.example.rezh.exceptions.DocumentNotFoundException;
+import com.example.rezh.exceptions.NewsNotFoundException;
 import com.example.rezh.repositories.DocumentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,11 +17,11 @@ public class DocumentService {
 
 
     public DocumentEntity getOneDocument(Integer id) throws DocumentNotFoundException {
-        DocumentEntity document = documentRepository.findById(id).get();
-        if (document == null){
-            throw new DocumentNotFoundException("Проект не найден");
+        var document = documentRepository.findById(id);
+        if (document.isEmpty()){
+            throw new DocumentNotFoundException("Документ не найден");
         }
-        return document;
+        return document.get();
     }
 
     public Iterable<DocumentEntity> getAllDocuments(){
@@ -31,12 +32,20 @@ public class DocumentService {
         return documentRepository.save(document);
     }
 
-    public Integer deleteDocument(Integer id) {
+    public Integer deleteDocument(Integer id) throws DocumentNotFoundException{
+        var document = documentRepository.findById(id);
+        if (document.isEmpty()){
+            throw new DocumentNotFoundException("Документ не найден");
+        }
         documentRepository.deleteById(id);
         return id;
     }
 
     public void editDocument(DocumentEntity newDocument, Integer id) throws DocumentNotFoundException {
+        var document = documentRepository.findById(id);
+        if (document.isEmpty()){
+            throw new DocumentNotFoundException("Документ не найден");
+        }
         DocumentEntity currentDocument = getOneDocument(id);
         if (newDocument.getTitle() != null)
             currentDocument.setTitle(newDocument.getTitle());
